Removed unneeded DAO implementations and moved over UserDAO companion object
