--- app/models/daos/PasswordInfoDAO.scala
@@ -1,38 +1,0 @@
-package models.daos
-
-import com.mohiva.play.silhouette.api.LoginInfo
-import com.mohiva.play.silhouette.api.util.PasswordInfo
-import com.mohiva.play.silhouette.impl.daos.DelegableAuthInfoDAO
-import models.daos.PasswordInfoDAO._
-
-import scala.collection.mutable
-import scala.concurrent.Future
-
-class PasswordInfoDAO extends DelegableAuthInfoDAO[PasswordInfo] {
-
-  /**
-   * Saves the password info.
-   *
-   * @param loginInfo The login info for which the auth info should be saved.
-   * @param authInfo The password info to save.
-   * @return The saved password info.
-   */
-  def save(loginInfo: LoginInfo, authInfo: PasswordInfo): Future[PasswordInfo] = {
-    data += (loginInfo -> authInfo)
-    Future.successful(authInfo)
-  }
-
-  /**
-   * Finds the password info which is linked with the specified login info.
-   *
-   * @param loginInfo The linked login info.
-   * @return The retrieved password info or None if no password info could be retrieved for the given login info.
-   */
-  def find(loginInfo: LoginInfo): Future[Option[PasswordInfo]] = {
-    Future.successful(data.get(loginInfo))
-  }
-}
-
-object PasswordInfoDAO {
-  var data: mutable.HashMap[LoginInfo, PasswordInfo] = mutable.HashMap()
-}
